docker/roadmap-ocr/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pillow==11.3.0
pytesseract==0.3.13
